Refactor bookmark command and AnimeData
