+++ src/main/java/seedu/duke/Duke.java
@@ -2,6 +2,7 @@ package seedu.duke;
 
 import seedu.duke.anime.Anime;
 import seedu.duke.anime.AnimeData;
+import seedu.duke.anime.AnimeStorage;
 import seedu.duke.bookmark.Bookmark;
 import seedu.duke.command.Command;
 import seedu.duke.exception.AniException;
@@ -11,6 +12,7 @@ import seedu.duke.storage.Storage;
 import seedu.duke.ui.Ui;
 import seedu.duke.watchlist.Watchlist;
 
+import java.io.IOException;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -19,28 +21,38 @@ import java.util.Scanner;
 public class Duke {
     private static final String USER_PROFILE_FILE_NAME = "userprofile.txt";
     private static final String WATCHLIST_FILE_NAME = "watchlist.txt";
+    private static final String ANIME_DATA_SOURCE_FOLDER = "/data/AniListData";
 
-    private final Ui ui;
-    private final Parser parser;
-    private final Storage storage;
+    private Ui ui;
+    private Parser parser;
+    private Storage storage;
+    private AnimeStorage animeStorage;
     private UserProfile userProfile;
-    private final Watchlist currentWatchlist;
-    private final ArrayList<Watchlist> watchlists;
+    private Watchlist currentWatchlist;
+    private ArrayList<Watchlist> watchlists;
+    private AnimeData animeData;
+    private Bookmark bookmark;
 
     public Duke() {
         ui = new Ui();
+        ui.printWelcomeMessage();
         parser = new Parser();
         storage = new Storage(USER_PROFILE_FILE_NAME, WATCHLIST_FILE_NAME);
-
-        ui.printWelcomeMessage();
         userProfile = storage.readUserProfileFile(ui);
         watchlists = storage.readWatchlistFile(ui);
-
+        bookmark = new Bookmark();
+        try {
+            animeStorage = new AnimeStorage(ANIME_DATA_SOURCE_FOLDER);
+            animeData = new AnimeData(animeStorage.readAnimeDatabase());
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
         if (watchlists.isEmpty()) {
             currentWatchlist = new Watchlist("Default");
         } else {
             currentWatchlist = watchlists.get(0);
         }
+
     }
 
     public void run() {
@@ -55,7 +67,7 @@ public class Duke {
                 command = Parser.getCommand(userInput);
                 // now passing in many parameters into execute, 
                 // but maybe can reduce in the future after refactoring?
-                command.execute(ui, storage, currentWatchlist, watchlists);
+                command.execute(ui, storage, animeData, currentWatchlist, watchlists, bookmark);
             } catch (AniException exception) {
                 ui.printErrorMessage(exception.getMessage());
             }
@@ -69,8 +81,7 @@ public class Duke {
     // TODO: Organize the methods below to their respective class.
     // TEMPORARY, REMOVE ALL BELOW WHEN DONE REFACTORING!
 
-    private static AnimeData animeData;
-    private static Bookmark bookmark;
+
     private static final Scanner CONSOLE = new Scanner(System.in);
 
     private UserProfile quickStart() {
@@ -221,80 +232,6 @@ public class Duke {
         }
     }
 
-    /**
-     * Bookmarks an anime.
-     */
-    private void bookmarkAnime(String description) {
-        if (description.contains(" ")) {
-            String[] descriptionSplit = description.split(" ", 2);
-            // Code to be added
-            String commandOption = descriptionSplit[0];
-            String commandArgument = descriptionSplit[1];
-            if (commandOption.equals("-a")) {
-                if (isInteger(commandArgument)) {
-                    int animeDataListIndex = Integer.parseInt(commandArgument);
-                    Anime anime = animeData.getAnimeByID(animeDataListIndex);
-                    System.out.println("Saving " + anime.getAnimeID() + ". "
-                            + anime.getAnimeName() + " " + " to bookmark.");
-                    bookmark.addAnimeBookmark(anime.getAnimeName());
-                } else {
-                    ArrayList<Anime> findList = animeData.findName(commandArgument);
-                    for (Anime anime : findList) {
-                        System.out.println("\t" + anime.getAnimeID() + ". " + anime.getAnimeName());
-                    }
-                }
-            } else if (commandOption.equals("-d")) {
-                int bookmarkIndex = Integer.parseInt(commandArgument);
-                String animeName = bookmark.getAnimeBookmarkByIndex(bookmarkIndex - 1);
-                System.out.println("Removing " + animeName + "! :(");
-                bookmark.removeAnimeBookmark(bookmarkIndex - 1);
-            } else {
-                int bookmarkIndex = Integer.parseInt(commandOption);
-                String[] commandArgumentSplit = commandArgument.split(" ", 2);
-                String commandOption2 = commandArgumentSplit[0];
-                String commandArgument2 = commandArgumentSplit[1];
-                if (commandOption2.equals("-e")) {
-                    int episode = Integer.parseInt(commandArgument2);
-                    bookmark.editAnimeBookmarkEpisode(bookmarkIndex - 1, episode);
-                    String animeName = bookmark.getAnimeBookmarkByIndex(bookmarkIndex - 1);
-                    System.out.println("Editing " + animeName + " to have " + episode + " episode");
-                }
-            }
-        } else {
-            if (description.equals("-l")) {
-                String bookmarks = bookmark.animeListInString();
-                System.out.println(bookmarks);
-            }
-        }
-    }
-
-    /**
-     * Check if input string provided is a valid string.
-     * False if string contains alpha, null and 0 length
-     * @param str : String to check
-     * @return
-     */
-    public boolean isInteger(String str) {
-        int length = str.length();
-        if (length == 0) {
-            return false;
-        }
-        int i = 0;
-        if (str.charAt(0) == '-') {
-            if (length == 1) {
-                return false;
-            }
-            i = 1;
-        }
-        for (; i < length; i++) {
-            char c = str.charAt(i);
-            if (c < '0' || c > '9') {
-                return false;
-            }
-        }
-        return true;
-    }
-    
     //Sample Usage of AnimeList Class [To Be Deleted]
     private void createAnimeList() {
         System.out.println("===Running Sample Anime List Class===");
