+++ src/main/java/seedu/duke/command/CreateWatchlistCommand.java
@@ -1,5 +1,7 @@
 package seedu.duke.command;
 
+import seedu.duke.anime.AnimeData;
+import seedu.duke.bookmark.Bookmark;
 import seedu.duke.storage.Storage;
 import seedu.duke.ui.Ui;
 import seedu.duke.watchlist.Watchlist;
@@ -13,8 +15,8 @@ public class CreateWatchlistCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, Storage storage, Watchlist currentWatchlist,
-                        ArrayList<Watchlist> watchlists) {
+    public void execute(Ui ui, Storage storage, AnimeData animeData, Watchlist currentWatchlist,
+                        ArrayList<Watchlist> watchlists, Bookmark bookmark) {
         String[] descriptionSplit = description.split(" ", 2);
         String commandOption = descriptionSplit[0];
         String watchlistName = descriptionSplit[1];
