+++ src/main/java/seedu/duke/command/HelpCommand.java
@@ -1,5 +1,7 @@
 package seedu.duke.command;
 
+import seedu.duke.anime.AnimeData;
+import seedu.duke.bookmark.Bookmark;
 import seedu.duke.storage.Storage;
 import seedu.duke.ui.Ui;
 import seedu.duke.watchlist.Watchlist;
@@ -16,8 +18,8 @@ public class HelpCommand extends Command {
      * Shows help function.
      */
     @Override
-    public void execute(Ui ui, Storage storage, Watchlist currentWatchlist,
-                        ArrayList<Watchlist> watchlists) {
+    public void execute(Ui ui, Storage storage, AnimeData animeData, Watchlist currentWatchlist,
+                        ArrayList<Watchlist> watchlists, Bookmark bookmark) {
         // Code to be added
 
         // Print for testing
